requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
schedule==1.2.0
python-dotenv==1.0.0
tqdm==4.66.1
fake-useragent==1.3.0
pymongo==4.5.0
pandas==2.0.3
selenium==4.15.0
webdriver-manager==4.0.1
playwright==1.38.0 